+++ CHANGELOG.md
@@ -28,6 +28,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Expanded the data source protocol to also include async fetch functions for async
   data sources
 - Updated StormCast coords to be HRRR index, output coords still provide lat lon
+- Interpolation AFNO model load_model now accepts prognostic model
 
 ### Deprecated